% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CmdTools.R, R/zzz.R
\name{CmdTools}
\alias{CmdTools}
\alias{cmd_tools}
\title{CmdTools class}
\usage{
cmd_tools()
}
\description{
Some description
}
\details{
Some details.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CmdTools-new}{\code{CmdTools$new()}}
\item \href{#method-CmdTools-ipynb_to_py}{\code{CmdTools$ipynb_to_py()}}
\item \href{#method-CmdTools-ipynb_strip_output}{\code{CmdTools$ipynb_strip_output()}}
\item \href{#method-CmdTools-clone}{\code{CmdTools$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CmdTools-new"></a>}}
\if{latex}{\out{\hypertarget{method-CmdTools-new}{}}}
\subsection{Method \code{new()}}{
CmdTools class constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CmdTools$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CmdTools-ipynb_to_py"></a>}}
\if{latex}{\out{\hypertarget{method-CmdTools-ipynb_to_py}{}}}
\subsection{Method \code{ipynb_to_py()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CmdTools$ipynb_to_py(
  paths,
  suppress_messages = T,
  ignore.stdout = T,
  ignore.stderr = T,
  remove_comments = F,
  remove_non_code = F,
  remove_output = F
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CmdTools-ipynb_strip_output"></a>}}
\if{latex}{\out{\hypertarget{method-CmdTools-ipynb_strip_output}{}}}
\subsection{Method \code{ipynb_strip_output()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CmdTools$ipynb_strip_output(
  paths,
  suppress_messages = T,
  ignore.stdout = T,
  ignore.stderr = T
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CmdTools-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CmdTools-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CmdTools$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
