% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGenerator.R
\name{DataGenerator}
\alias{DataGenerator}
\title{DataGenerator class}
\description{
Some description
}
\details{
Some details.
}
\note{
For debugging only. Don't you dare use this in a package!
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataGenerator-expose}{\code{DataGenerator$expose()}}
\item \href{#method-DataGenerator-new}{\code{DataGenerator$new()}}
\item \href{#method-DataGenerator-sample}{\code{DataGenerator$sample()}}
\item \href{#method-DataGenerator-get_last_data}{\code{DataGenerator$get_last_data()}}
\item \href{#method-DataGenerator-clone}{\code{DataGenerator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataGenerator-expose"></a>}}
\if{latex}{\out{\hypertarget{method-DataGenerator-expose}{}}}
\subsection{Method \code{expose()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataGenerator$expose()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataGenerator-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataGenerator-new}{}}}
\subsection{Method \code{new()}}{
RaiseModel class constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataGenerator$new(study_design)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataGenerator-sample"></a>}}
\if{latex}{\out{\hypertarget{method-DataGenerator-sample}{}}}
\subsection{Method \code{sample()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataGenerator$sample(labeled = T, centering = F, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataGenerator-get_last_data"></a>}}
\if{latex}{\out{\hypertarget{method-DataGenerator-get_last_data}{}}}
\subsection{Method \code{get_last_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataGenerator$get_last_data(labeled = T)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataGenerator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataGenerator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataGenerator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
