% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exporter.R
\name{Exporter}
\alias{Exporter}
\title{Exporter class}
\description{
Some description
}
\details{
Some details.
}
\note{
For debugging only. Don't you dare use this in a package!
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Exporter-expose}{\code{Exporter$expose()}}
\item \href{#method-Exporter-new}{\code{Exporter$new()}}
\item \href{#method-Exporter-export}{\code{Exporter$export()}}
\item \href{#method-Exporter-clone}{\code{Exporter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Exporter-expose"></a>}}
\if{latex}{\out{\hypertarget{method-Exporter-expose}{}}}
\subsection{Method \code{expose()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exporter$expose()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Exporter-new"></a>}}
\if{latex}{\out{\hypertarget{method-Exporter-new}{}}}
\subsection{Method \code{new()}}{
Exporter class constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exporter$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Exporter-export"></a>}}
\if{latex}{\out{\hypertarget{method-Exporter-export}{}}}
\subsection{Method \code{export()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exporter$export(
  output_file = "summary.html",
  output_dir = NULL,
  open = T,
  keep_files = F,
  export = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Exporter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Exporter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exporter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
