% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExporterQuiz.R
\name{ExporterQuiz}
\alias{ExporterQuiz}
\title{ExporterQuiz class}
\description{
Some description
}
\details{
Some details.
}
\note{
For debugging only. Don't you dare use this in a package!
}
\section{Super class}{
\code{\link[raise:Exporter]{raise::Exporter}} -> \code{ExporterQuiz}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ExporterQuiz-expose}{\code{ExporterQuiz$expose()}}
\item \href{#method-ExporterQuiz-new}{\code{ExporterQuiz$new()}}
\item \href{#method-ExporterQuiz-export}{\code{ExporterQuiz$export()}}
\item \href{#method-ExporterQuiz-export_exams}{\code{ExporterQuiz$export_exams()}}
\item \href{#method-ExporterQuiz-clone}{\code{ExporterQuiz$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExporterQuiz-expose"></a>}}
\if{latex}{\out{\hypertarget{method-ExporterQuiz-expose}{}}}
\subsection{Method \code{expose()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExporterQuiz$expose()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExporterQuiz-new"></a>}}
\if{latex}{\out{\hypertarget{method-ExporterQuiz-new}{}}}
\subsection{Method \code{new()}}{
ExporterQuiz class constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExporterQuiz$new(quiz)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExporterQuiz-export"></a>}}
\if{latex}{\out{\hypertarget{method-ExporterQuiz-export}{}}}
\subsection{Method \code{export()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExporterQuiz$export(engine = "exams", ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExporterQuiz-export_exams"></a>}}
\if{latex}{\out{\hypertarget{method-ExporterQuiz-export_exams}{}}}
\subsection{Method \code{export_exams()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExporterQuiz$export_exams(
  output_file = "summary.html",
  output_dir = NULL,
  destination = "pdf",
  open = T,
  keep_files = F,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExporterQuiz-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ExporterQuiz-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExporterQuiz$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
