% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExporterStudy.R, R/ExporterStudyBinary.R,
%   R/ExporterStudyGaussian.R
\name{ExporterStudy}
\alias{ExporterStudy}
\alias{ExporterStudyBinary}
\alias{ExporterStudyGaussian}
\title{ExporterStudy class}
\description{
Some description

Some description

Some description
}
\details{
Some details.

Some details.

Some details.
}
\note{
For debugging only. Don't you dare use this in a package!
}
\section{Super class}{
\code{\link[raise:Exporter]{raise::Exporter}} -> \code{ExporterStudy}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ExporterStudy-expose}{\code{ExporterStudy$expose()}}
\item \href{#method-ExporterStudy-new}{\code{ExporterStudy$new()}}
\item \href{#method-ExporterStudy-export}{\code{ExporterStudy$export()}}
\item \href{#method-ExporterStudy-clone}{\code{ExporterStudy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExporterStudy-expose"></a>}}
\if{latex}{\out{\hypertarget{method-ExporterStudy-expose}{}}}
\subsection{Method \code{expose()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExporterStudy$expose()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExporterStudy-new"></a>}}
\if{latex}{\out{\hypertarget{method-ExporterStudy-new}{}}}
\subsection{Method \code{new()}}{
ExporterStudy class constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExporterStudy$new(study_design)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExporterStudy-export"></a>}}
\if{latex}{\out{\hypertarget{method-ExporterStudy-export}{}}}
\subsection{Method \code{export()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExporterStudy$export(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExporterStudy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ExporterStudy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExporterStudy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super classes}{
\code{\link[raise:Exporter]{raise::Exporter}} -> \code{\link[raise:ExporterStudy]{raise::ExporterStudy}} -> \code{ExporterStudyBinary}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ExporterStudyBinary-new}{\code{ExporterStudyBinary$new()}}
\item \href{#method-ExporterStudyBinary-clone}{\code{ExporterStudyBinary$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="raise" data-topic="ExporterStudy" data-id="export"><a href='../../raise/html/ExporterStudy.html#method-ExporterStudy-export'><code>raise::ExporterStudy$export()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="ExporterStudy" data-id="expose"><a href='../../raise/html/ExporterStudy.html#method-ExporterStudy-expose'><code>raise::ExporterStudy$expose()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExporterStudyBinary-new"></a>}}
\if{latex}{\out{\hypertarget{method-ExporterStudyBinary-new}{}}}
\subsection{Method \code{new()}}{
ExporterStudyBinary class constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExporterStudyBinary$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExporterStudyBinary-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ExporterStudyBinary-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExporterStudyBinary$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super classes}{
\code{\link[raise:Exporter]{raise::Exporter}} -> \code{\link[raise:ExporterStudy]{raise::ExporterStudy}} -> \code{ExporterStudyGaussian}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ExporterStudyGaussian-new}{\code{ExporterStudyGaussian$new()}}
\item \href{#method-ExporterStudyGaussian-clone}{\code{ExporterStudyGaussian$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="raise" data-topic="ExporterStudy" data-id="export"><a href='../../raise/html/ExporterStudy.html#method-ExporterStudy-export'><code>raise::ExporterStudy$export()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="ExporterStudy" data-id="expose"><a href='../../raise/html/ExporterStudy.html#method-ExporterStudy-expose'><code>raise::ExporterStudy$expose()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExporterStudyGaussian-new"></a>}}
\if{latex}{\out{\hypertarget{method-ExporterStudyGaussian-new}{}}}
\subsection{Method \code{new()}}{
ExporterStudyGaussian class constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExporterStudyGaussian$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExporterStudyGaussian-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ExporterStudyGaussian-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExporterStudyGaussian$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
