% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ai.R
\name{extract_ai}
\alias{extract_ai}
\title{Parse a string or vector to a tibble}
\usage{
extract_ai(text, cols, ...)
}
\arguments{
\item{cols}{The columns of the tibble to be returned. Could be something like `c("State", "Area", "Population")`}

\item{prompt}{The prompt you'd like to run. It could be something like "A spreadsheet of all 50 US states"}
}
\description{
You will need an OpenAI API key.
}
\examples{
us_states <- create_ai("All 50 US states", c("State", "Area", "Population"))
mountains <- create_ai("Top 5 tallest mountains", c("Mountain", "Country", "Height in metres"))
peppers <- create_ai("Top 10 hottest peppers", c("Pepper", "Country of origin", "Scoville heat units"))
}
