% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gpt.R, R/zzz.R
\name{Gpt}
\alias{Gpt}
\alias{set_gpt_model}
\alias{get_gpt_model}
\alias{set_api_key}
\alias{power_up}
\alias{get_api_key}
\alias{gpt}
\title{Gpt class}
\usage{
set_gpt_model(model)

get_gpt_model()

set_api_key(api_key)

power_up(api_key)

get_api_key()

gpt()
}
\arguments{
\item{api_key}{GPT API key. Must be set before first use.}

\item{version}{GPT version. Default is "gpt-4o".}
}
\description{
Some description
}
\details{
Some details.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Gpt-new}{\code{Gpt$new()}}
\item \href{#method-Gpt-is_ready}{\code{Gpt$is_ready()}}
\item \href{#method-Gpt-set_api_key}{\code{Gpt$set_api_key()}}
\item \href{#method-Gpt-reconnect}{\code{Gpt$reconnect()}}
\item \href{#method-Gpt-check_assistant_ready}{\code{Gpt$check_assistant_ready()}}
\item \href{#method-Gpt-check_client_ready}{\code{Gpt$check_client_ready()}}
\item \href{#method-Gpt-power_up}{\code{Gpt$power_up()}}
\item \href{#method-Gpt-shut_down}{\code{Gpt$shut_down()}}
\item \href{#method-Gpt-get_assistant_id}{\code{Gpt$get_assistant_id()}}
\item \href{#method-Gpt-get_assistant_by_name}{\code{Gpt$get_assistant_by_name()}}
\item \href{#method-Gpt-create_assistants}{\code{Gpt$create_assistants()}}
\item \href{#method-Gpt-delete_assistant}{\code{Gpt$delete_assistant()}}
\item \href{#method-Gpt-create_thread}{\code{Gpt$create_thread()}}
\item \href{#method-Gpt-delete_thread}{\code{Gpt$delete_thread()}}
\item \href{#method-Gpt-list_messages}{\code{Gpt$list_messages()}}
\item \href{#method-Gpt-flush_assistants}{\code{Gpt$flush_assistants()}}
\item \href{#method-Gpt-create_vector_store}{\code{Gpt$create_vector_store()}}
\item \href{#method-Gpt-add_vector_stores_to_thread}{\code{Gpt$add_vector_stores_to_thread()}}
\item \href{#method-Gpt-delete_vector_store}{\code{Gpt$delete_vector_store()}}
\item \href{#method-Gpt-list_vector_stores}{\code{Gpt$list_vector_stores()}}
\item \href{#method-Gpt-create_message}{\code{Gpt$create_message()}}
\item \href{#method-Gpt-delete_message}{\code{Gpt$delete_message()}}
\item \href{#method-Gpt-flush_vector_stores}{\code{Gpt$flush_vector_stores()}}
\item \href{#method-Gpt-list_files}{\code{Gpt$list_files()}}
\item \href{#method-Gpt-flush_files}{\code{Gpt$flush_files()}}
\item \href{#method-Gpt-delete_file}{\code{Gpt$delete_file()}}
\item \href{#method-Gpt-delete_files}{\code{Gpt$delete_files()}}
\item \href{#method-Gpt-upload_files}{\code{Gpt$upload_files()}}
\item \href{#method-Gpt-upload_files_to_vector_store}{\code{Gpt$upload_files_to_vector_store()}}
\item \href{#method-Gpt-delete_files_from_vector_store}{\code{Gpt$delete_files_from_vector_store()}}
\item \href{#method-Gpt-completion_request}{\code{Gpt$completion_request()}}
\item \href{#method-Gpt-verify_json}{\code{Gpt$verify_json()}}
\item \href{#method-Gpt-assistant_request}{\code{Gpt$assistant_request()}}
\item \href{#method-Gpt-clone}{\code{Gpt$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gpt-new"></a>}}
\if{latex}{\out{\hypertarget{method-Gpt-new}{}}}
\subsection{Method \code{new()}}{
Gpt class constructor.
Checks whether python and module openai are installed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gpt$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gpt-is_ready"></a>}}
\if{latex}{\out{\hypertarget{method-Gpt-is_ready}{}}}
\subsection{Method \code{is_ready()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gpt$is_ready(ask = T)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ask}}{Logical. Default is FALSE. Indicates whether user is asked whether
openai should be installed if it is not available.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gpt-set_api_key"></a>}}
\if{latex}{\out{\hypertarget{method-Gpt-set_api_key}{}}}
\subsection{Method \code{set_api_key()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gpt$set_api_key(api_key)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gpt-reconnect"></a>}}
\if{latex}{\out{\hypertarget{method-Gpt-reconnect}{}}}
\subsection{Method \code{reconnect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gpt$reconnect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gpt-check_assistant_ready"></a>}}
\if{latex}{\out{\hypertarget{method-Gpt-check_assistant_ready}{}}}
\subsection{Method \code{check_assistant_ready()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gpt$check_assistant_ready()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gpt-check_client_ready"></a>}}
\if{latex}{\out{\hypertarget{method-Gpt-check_client_ready}{}}}
\subsection{Method \code{check_client_ready()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gpt$check_client_ready()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gpt-power_up"></a>}}
\if{latex}{\out{\hypertarget{method-Gpt-power_up}{}}}
\subsection{Method \code{power_up()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gpt$power_up(api_key, reuse = TRUE, import_main = TRUE, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gpt-shut_down"></a>}}
\if{latex}{\out{\hypertarget{method-Gpt-shut_down}{}}}
\subsection{Method \code{shut_down()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gpt$shut_down()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gpt-get_assistant_id"></a>}}
\if{latex}{\out{\hypertarget{method-Gpt-get_assistant_id}{}}}
\subsection{Method \code{get_assistant_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gpt$get_assistant_id(assistant_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gpt-get_assistant_by_name"></a>}}
\if{latex}{\out{\hypertarget{method-Gpt-get_assistant_by_name}{}}}
\subsection{Method \code{get_assistant_by_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gpt$get_assistant_by_name(assistant_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gpt-create_assistants"></a>}}
\if{latex}{\out{\hypertarget{method-Gpt-create_assistants}{}}}
\subsection{Method \code{create_assistants()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gpt$create_assistants(reuse = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gpt-delete_assistant"></a>}}
\if{latex}{\out{\hypertarget{method-Gpt-delete_assistant}{}}}
\subsection{Method \code{delete_assistant()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gpt$delete_assistant(assistant_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gpt-create_thread"></a>}}
\if{latex}{\out{\hypertarget{method-Gpt-create_thread}{}}}
\subsection{Method \code{create_thread()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gpt$create_thread()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gpt-delete_thread"></a>}}
\if{latex}{\out{\hypertarget{method-Gpt-delete_thread}{}}}
\subsection{Method \code{delete_thread()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gpt$delete_thread(thread_id, delete_messages = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gpt-list_messages"></a>}}
\if{latex}{\out{\hypertarget{method-Gpt-list_messages}{}}}
\subsection{Method \code{list_messages()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gpt$list_messages(thread_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gpt-flush_assistants"></a>}}
\if{latex}{\out{\hypertarget{method-Gpt-flush_assistants}{}}}
\subsection{Method \code{flush_assistants()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gpt$flush_assistants()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gpt-create_vector_store"></a>}}
\if{latex}{\out{\hypertarget{method-Gpt-create_vector_store}{}}}
\subsection{Method \code{create_vector_store()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gpt$create_vector_store(thread_id = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gpt-add_vector_stores_to_thread"></a>}}
\if{latex}{\out{\hypertarget{method-Gpt-add_vector_stores_to_thread}{}}}
\subsection{Method \code{add_vector_stores_to_thread()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gpt$add_vector_stores_to_thread(vector_store_ids, thread_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gpt-delete_vector_store"></a>}}
\if{latex}{\out{\hypertarget{method-Gpt-delete_vector_store}{}}}
\subsection{Method \code{delete_vector_store()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gpt$delete_vector_store(vector_store_id, delete_files = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gpt-list_vector_stores"></a>}}
\if{latex}{\out{\hypertarget{method-Gpt-list_vector_stores}{}}}
\subsection{Method \code{list_vector_stores()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gpt$list_vector_stores(names_only = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gpt-create_message"></a>}}
\if{latex}{\out{\hypertarget{method-Gpt-create_message}{}}}
\subsection{Method \code{create_message()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gpt$create_message(thread_id, role, content)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gpt-delete_message"></a>}}
\if{latex}{\out{\hypertarget{method-Gpt-delete_message}{}}}
\subsection{Method \code{delete_message()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gpt$delete_message(thread_id, message_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gpt-flush_vector_stores"></a>}}
\if{latex}{\out{\hypertarget{method-Gpt-flush_vector_stores}{}}}
\subsection{Method \code{flush_vector_stores()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gpt$flush_vector_stores(delete_files = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gpt-list_files"></a>}}
\if{latex}{\out{\hypertarget{method-Gpt-list_files}{}}}
\subsection{Method \code{list_files()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gpt$list_files()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gpt-flush_files"></a>}}
\if{latex}{\out{\hypertarget{method-Gpt-flush_files}{}}}
\subsection{Method \code{flush_files()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gpt$flush_files()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gpt-delete_file"></a>}}
\if{latex}{\out{\hypertarget{method-Gpt-delete_file}{}}}
\subsection{Method \code{delete_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gpt$delete_file(file_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gpt-delete_files"></a>}}
\if{latex}{\out{\hypertarget{method-Gpt-delete_files}{}}}
\subsection{Method \code{delete_files()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gpt$delete_files(file_ids)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gpt-upload_files"></a>}}
\if{latex}{\out{\hypertarget{method-Gpt-upload_files}{}}}
\subsection{Method \code{upload_files()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gpt$upload_files(file_paths)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gpt-upload_files_to_vector_store"></a>}}
\if{latex}{\out{\hypertarget{method-Gpt-upload_files_to_vector_store}{}}}
\subsection{Method \code{upload_files_to_vector_store()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gpt$upload_files_to_vector_store(file_paths, vector_store_id = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gpt-delete_files_from_vector_store"></a>}}
\if{latex}{\out{\hypertarget{method-Gpt-delete_files_from_vector_store}{}}}
\subsection{Method \code{delete_files_from_vector_store()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gpt$delete_files_from_vector_store(file_ids, vector_store_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gpt-completion_request"></a>}}
\if{latex}{\out{\hypertarget{method-Gpt-completion_request}{}}}
\subsection{Method \code{completion_request()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gpt$completion_request(messages)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gpt-verify_json"></a>}}
\if{latex}{\out{\hypertarget{method-Gpt-verify_json}{}}}
\subsection{Method \code{verify_json()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gpt$verify_json(json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gpt-assistant_request"></a>}}
\if{latex}{\out{\hypertarget{method-Gpt-assistant_request}{}}}
\subsection{Method \code{assistant_request()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gpt$assistant_request(
  thread_id,
  assistant_name = "assistant_json",
  verify_json = TRUE,
  drop_message = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gpt-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Gpt-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gpt$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
