% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GptThread.R
\name{GptThread}
\alias{GptThread}
\title{GptThread class}
\description{
Some description
}
\details{
Some details.
}
\note{
For debugging only. Don't you dare use this in a package!
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GptThread-copy_from}{\code{GptThread$copy_from()}}
\item \href{#method-GptThread-expose}{\code{GptThread$expose()}}
\item \href{#method-GptThread-new}{\code{GptThread$new()}}
\item \href{#method-GptThread-create_vector_store}{\code{GptThread$create_vector_store()}}
\item \href{#method-GptThread-power_up}{\code{GptThread$power_up()}}
\item \href{#method-GptThread-cover_tracks}{\code{GptThread$cover_tracks()}}
\item \href{#method-GptThread-upload_files}{\code{GptThread$upload_files()}}
\item \href{#method-GptThread-parse_files}{\code{GptThread$parse_files()}}
\item \href{#method-GptThread-chat}{\code{GptThread$chat()}}
\item \href{#method-GptThread-get_messages}{\code{GptThread$get_messages()}}
\item \href{#method-GptThread-get_last_message}{\code{GptThread$get_last_message()}}
\item \href{#method-GptThread-restore}{\code{GptThread$restore()}}
\item \href{#method-GptThread-clone}{\code{GptThread$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GptThread-copy_from"></a>}}
\if{latex}{\out{\hypertarget{method-GptThread-copy_from}{}}}
\subsection{Method \code{copy_from()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GptThread$copy_from(from, cache = T)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GptThread-expose"></a>}}
\if{latex}{\out{\hypertarget{method-GptThread-expose}{}}}
\subsection{Method \code{expose()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GptThread$expose()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GptThread-new"></a>}}
\if{latex}{\out{\hypertarget{method-GptThread-new}{}}}
\subsection{Method \code{new()}}{
GptThread class constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GptThread$new(gpt_thread = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GptThread-create_vector_store"></a>}}
\if{latex}{\out{\hypertarget{method-GptThread-create_vector_store}{}}}
\subsection{Method \code{create_vector_store()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GptThread$create_vector_store()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GptThread-power_up"></a>}}
\if{latex}{\out{\hypertarget{method-GptThread-power_up}{}}}
\subsection{Method \code{power_up()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GptThread$power_up(force = F)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GptThread-cover_tracks"></a>}}
\if{latex}{\out{\hypertarget{method-GptThread-cover_tracks}{}}}
\subsection{Method \code{cover_tracks()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GptThread$cover_tracks()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GptThread-upload_files"></a>}}
\if{latex}{\out{\hypertarget{method-GptThread-upload_files}{}}}
\subsection{Method \code{upload_files()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GptThread$upload_files(file_paths)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GptThread-parse_files"></a>}}
\if{latex}{\out{\hypertarget{method-GptThread-parse_files}{}}}
\subsection{Method \code{parse_files()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GptThread$parse_files(paths)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GptThread-chat"></a>}}
\if{latex}{\out{\hypertarget{method-GptThread-chat}{}}}
\subsection{Method \code{chat()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GptThread$chat(message, respond = T, print = T, verify_json = F)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GptThread-get_messages"></a>}}
\if{latex}{\out{\hypertarget{method-GptThread-get_messages}{}}}
\subsection{Method \code{get_messages()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GptThread$get_messages(force = F)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GptThread-get_last_message"></a>}}
\if{latex}{\out{\hypertarget{method-GptThread-get_last_message}{}}}
\subsection{Method \code{get_last_message()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GptThread$get_last_message(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GptThread-restore"></a>}}
\if{latex}{\out{\hypertarget{method-GptThread-restore}{}}}
\subsection{Method \code{restore()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GptThread$restore(gpt_thread = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GptThread-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GptThread-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GptThread$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
