% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PdfExtractor.R, R/zzz.R
\name{PdfExtractor}
\alias{PdfExtractor}
\alias{pdf_extractor}
\title{PdfExtractor class}
\usage{
pdf_extractor()
}
\description{
Some description
}
\details{
Some details.
}
\keyword{internal}
\section{Super class}{
\code{\link[raise:PythonInterface]{raise::PythonInterface}} -> \code{PdfExtractor}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PdfExtractor-new}{\code{PdfExtractor$new()}}
\item \href{#method-PdfExtractor-extract}{\code{PdfExtractor$extract()}}
\item \href{#method-PdfExtractor-clone}{\code{PdfExtractor$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="raise" data-topic="PythonInterface" data-id="is_ready"><a href='../../raise/html/PythonInterface.html#method-PythonInterface-is_ready'><code>raise::PythonInterface$is_ready()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PdfExtractor-new"></a>}}
\if{latex}{\out{\hypertarget{method-PdfExtractor-new}{}}}
\subsection{Method \code{new()}}{
PdfExtractor class constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PdfExtractor$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PdfExtractor-extract"></a>}}
\if{latex}{\out{\hypertarget{method-PdfExtractor-extract}{}}}
\subsection{Method \code{extract()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PdfExtractor$extract(paths)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PdfExtractor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PdfExtractor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PdfExtractor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
