% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PdfExtractor_backup.R
\name{PdfExtractorBackup}
\alias{PdfExtractorBackup}
\title{PdfExtractorBackup class}
\description{
Some description
}
\details{
Some details.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PdfExtractorBackup-new}{\code{PdfExtractorBackup$new()}}
\item \href{#method-PdfExtractorBackup-is_ready}{\code{PdfExtractorBackup$is_ready()}}
\item \href{#method-PdfExtractorBackup-extract}{\code{PdfExtractorBackup$extract()}}
\item \href{#method-PdfExtractorBackup-clone}{\code{PdfExtractorBackup$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PdfExtractorBackup-new"></a>}}
\if{latex}{\out{\hypertarget{method-PdfExtractorBackup-new}{}}}
\subsection{Method \code{new()}}{
PdfExtractor class constructor.
Checks whether python and module openai are installed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PdfExtractorBackup$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PdfExtractorBackup-is_ready"></a>}}
\if{latex}{\out{\hypertarget{method-PdfExtractorBackup-is_ready}{}}}
\subsection{Method \code{is_ready()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PdfExtractorBackup$is_ready(install = T, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ask}}{Logical. Default is FALSE. Indicates whether user is asked whether
openai should be installed if it is not available.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PdfExtractorBackup-extract"></a>}}
\if{latex}{\out{\hypertarget{method-PdfExtractorBackup-extract}{}}}
\subsection{Method \code{extract()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PdfExtractorBackup$extract(paths)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PdfExtractorBackup-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PdfExtractorBackup-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PdfExtractorBackup$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
