% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PyNotebookExtractor.R, R/zzz.R
\name{PyNotebookExtractor}
\alias{PyNotebookExtractor}
\alias{py_notebook_extractor}
\title{PyNotebookExtractor class}
\usage{
py_notebook_extractor()
}
\description{
Some description
}
\details{
Some details.
}
\keyword{internal}
\section{Super class}{
\code{\link[raise:PythonInterface]{raise::PythonInterface}} -> \code{PyNotebookExtractor}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PyNotebookExtractor-new}{\code{PyNotebookExtractor$new()}}
\item \href{#method-PyNotebookExtractor-extract}{\code{PyNotebookExtractor$extract()}}
\item \href{#method-PyNotebookExtractor-clone}{\code{PyNotebookExtractor$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="raise" data-topic="PythonInterface" data-id="is_ready"><a href='../../raise/html/PythonInterface.html#method-PythonInterface-is_ready'><code>raise::PythonInterface$is_ready()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PyNotebookExtractor-new"></a>}}
\if{latex}{\out{\hypertarget{method-PyNotebookExtractor-new}{}}}
\subsection{Method \code{new()}}{
PdfExtractor class constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PyNotebookExtractor$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PyNotebookExtractor-extract"></a>}}
\if{latex}{\out{\hypertarget{method-PyNotebookExtractor-extract}{}}}
\subsection{Method \code{extract()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PyNotebookExtractor$extract(paths)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PyNotebookExtractor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PyNotebookExtractor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PyNotebookExtractor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
