% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PythonInterface.R
\name{PythonInterface}
\alias{PythonInterface}
\title{PythonInterface class}
\description{
Some description
}
\details{
Some details.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PythonInterface-new}{\code{PythonInterface$new()}}
\item \href{#method-PythonInterface-is_ready}{\code{PythonInterface$is_ready()}}
\item \href{#method-PythonInterface-clone}{\code{PythonInterface$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PythonInterface-new"></a>}}
\if{latex}{\out{\hypertarget{method-PythonInterface-new}{}}}
\subsection{Method \code{new()}}{
PythonInterface class constructor.
Checks whether python and module openai are installed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PythonInterface$new(py_sources = c(), py_modules = c())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PythonInterface-is_ready"></a>}}
\if{latex}{\out{\hypertarget{method-PythonInterface-is_ready}{}}}
\subsection{Method \code{is_ready()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PythonInterface$is_ready(install = T, source_py_files = F, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ask}}{Logical. Default is FALSE. Indicates whether user is asked whether
openai should be installed if it is not available.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PythonInterface-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PythonInterface-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PythonInterface$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
