% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{question_mc}
\alias{question_mc}
\alias{question_numeric}
\title{Helper function for adding a new multiple choice question to the quiz.}
\usage{
question_mc(details = NULL, n_questions = 1, n_choices, n_correct = 1)

question_numeric(details = NULL, n_questions = 1)
}
\description{
Helper function for adding a new multiple choice question to the quiz.

Helper function for adding a new numeric question to the quiz.
}
