% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quiz.R
\name{Quiz}
\alias{Quiz}
\title{Quiz class}
\description{
Some description
}
\details{
Some details.
}
\note{
For debugging only. Don't you dare use this in a package!
}
\section{Super class}{
\code{\link[raise:GptThread]{raise::GptThread}} -> \code{Quiz}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Quiz-copy_from}{\code{Quiz$copy_from()}}
\item \href{#method-Quiz-expose}{\code{Quiz$expose()}}
\item \href{#method-Quiz-power_up}{\code{Quiz$power_up()}}
\item \href{#method-Quiz-new}{\code{Quiz$new()}}
\item \href{#method-Quiz-add_questions_legacy}{\code{Quiz$add_questions_legacy()}}
\item \href{#method-Quiz-add_questions}{\code{Quiz$add_questions()}}
\item \href{#method-Quiz-get_questions}{\code{Quiz$get_questions()}}
\item \href{#method-Quiz-shuffle_choices}{\code{Quiz$shuffle_choices()}}
\item \href{#method-Quiz-shuffle_questions}{\code{Quiz$shuffle_questions()}}
\item \href{#method-Quiz-get_questions_md}{\code{Quiz$get_questions_md()}}
\item \href{#method-Quiz-delete_question}{\code{Quiz$delete_question()}}
\item \href{#method-Quiz-flush_question}{\code{Quiz$flush_question()}}
\item \href{#method-Quiz-export}{\code{Quiz$export()}}
\item \href{#method-Quiz-get_description}{\code{Quiz$get_description()}}
\item \href{#method-Quiz-set_description}{\code{Quiz$set_description()}}
\item \href{#method-Quiz-get_title}{\code{Quiz$get_title()}}
\item \href{#method-Quiz-set_title}{\code{Quiz$set_title()}}
\item \href{#method-Quiz-restore}{\code{Quiz$restore()}}
\item \href{#method-Quiz-clone}{\code{Quiz$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="chat"><a href='../../raise/html/GptThread.html#method-GptThread-chat'><code>raise::GptThread$chat()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="cover_tracks"><a href='../../raise/html/GptThread.html#method-GptThread-cover_tracks'><code>raise::GptThread$cover_tracks()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="create_vector_store"><a href='../../raise/html/GptThread.html#method-GptThread-create_vector_store'><code>raise::GptThread$create_vector_store()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="get_last_message"><a href='../../raise/html/GptThread.html#method-GptThread-get_last_message'><code>raise::GptThread$get_last_message()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="get_messages"><a href='../../raise/html/GptThread.html#method-GptThread-get_messages'><code>raise::GptThread$get_messages()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="parse_files"><a href='../../raise/html/GptThread.html#method-GptThread-parse_files'><code>raise::GptThread$parse_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="upload_files"><a href='../../raise/html/GptThread.html#method-GptThread-upload_files'><code>raise::GptThread$upload_files()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Quiz-copy_from"></a>}}
\if{latex}{\out{\hypertarget{method-Quiz-copy_from}{}}}
\subsection{Method \code{copy_from()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Quiz$copy_from(from, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Quiz-expose"></a>}}
\if{latex}{\out{\hypertarget{method-Quiz-expose}{}}}
\subsection{Method \code{expose()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Quiz$expose()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Quiz-power_up"></a>}}
\if{latex}{\out{\hypertarget{method-Quiz-power_up}{}}}
\subsection{Method \code{power_up()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Quiz$power_up(force = F)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Quiz-new"></a>}}
\if{latex}{\out{\hypertarget{method-Quiz-new}{}}}
\subsection{Method \code{new()}}{
Quiz class constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Quiz$new(title = "", notes = "", ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Quiz-add_questions_legacy"></a>}}
\if{latex}{\out{\hypertarget{method-Quiz-add_questions_legacy}{}}}
\subsection{Method \code{add_questions_legacy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Quiz$add_questions_legacy(..., question_list = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Quiz-add_questions"></a>}}
\if{latex}{\out{\hypertarget{method-Quiz-add_questions}{}}}
\subsection{Method \code{add_questions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Quiz$add_questions(..., question_list = list(), details = NULL, print = F)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Quiz-get_questions"></a>}}
\if{latex}{\out{\hypertarget{method-Quiz-get_questions}{}}}
\subsection{Method \code{get_questions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Quiz$get_questions(shuffle_questions = F, shuffle_choices = F, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Quiz-shuffle_choices"></a>}}
\if{latex}{\out{\hypertarget{method-Quiz-shuffle_choices}{}}}
\subsection{Method \code{shuffle_choices()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Quiz$shuffle_choices()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Quiz-shuffle_questions"></a>}}
\if{latex}{\out{\hypertarget{method-Quiz-shuffle_questions}{}}}
\subsection{Method \code{shuffle_questions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Quiz$shuffle_questions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Quiz-get_questions_md"></a>}}
\if{latex}{\out{\hypertarget{method-Quiz-get_questions_md}{}}}
\subsection{Method \code{get_questions_md()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Quiz$get_questions_md(heading_level = 1, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Quiz-delete_question"></a>}}
\if{latex}{\out{\hypertarget{method-Quiz-delete_question}{}}}
\subsection{Method \code{delete_question()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Quiz$delete_question(index)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Quiz-flush_question"></a>}}
\if{latex}{\out{\hypertarget{method-Quiz-flush_question}{}}}
\subsection{Method \code{flush_question()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Quiz$flush_question(index)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Quiz-export"></a>}}
\if{latex}{\out{\hypertarget{method-Quiz-export}{}}}
\subsection{Method \code{export()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Quiz$export(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Quiz-get_description"></a>}}
\if{latex}{\out{\hypertarget{method-Quiz-get_description}{}}}
\subsection{Method \code{get_description()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Quiz$get_description(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Quiz-set_description"></a>}}
\if{latex}{\out{\hypertarget{method-Quiz-set_description}{}}}
\subsection{Method \code{set_description()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Quiz$set_description(description)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Quiz-get_title"></a>}}
\if{latex}{\out{\hypertarget{method-Quiz-get_title}{}}}
\subsection{Method \code{get_title()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Quiz$get_title()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Quiz-set_title"></a>}}
\if{latex}{\out{\hypertarget{method-Quiz-set_title}{}}}
\subsection{Method \code{set_title()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Quiz$set_title(title)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Quiz-restore"></a>}}
\if{latex}{\out{\hypertarget{method-Quiz-restore}{}}}
\subsection{Method \code{restore()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Quiz$restore(gpt_thread = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Quiz-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Quiz-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Quiz$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
