% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuizSlides.R
\name{QuizSlides}
\alias{QuizSlides}
\title{QuizSlides class}
\description{
Some description
}
\details{
Some details.
}
\section{Super classes}{
\code{\link[raise:GptThread]{raise::GptThread}} -> \code{\link[raise:Quiz]{raise::Quiz}} -> \code{QuizSlides}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-QuizSlides-new}{\code{QuizSlides$new()}}
\item \href{#method-QuizSlides-get_description}{\code{QuizSlides$get_description()}}
\item \href{#method-QuizSlides-clone}{\code{QuizSlides$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="chat"><a href='../../raise/html/GptThread.html#method-GptThread-chat'><code>raise::GptThread$chat()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="cover_tracks"><a href='../../raise/html/GptThread.html#method-GptThread-cover_tracks'><code>raise::GptThread$cover_tracks()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="create_vector_store"><a href='../../raise/html/GptThread.html#method-GptThread-create_vector_store'><code>raise::GptThread$create_vector_store()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="get_last_message"><a href='../../raise/html/GptThread.html#method-GptThread-get_last_message'><code>raise::GptThread$get_last_message()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="get_messages"><a href='../../raise/html/GptThread.html#method-GptThread-get_messages'><code>raise::GptThread$get_messages()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="parse_files"><a href='../../raise/html/GptThread.html#method-GptThread-parse_files'><code>raise::GptThread$parse_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="upload_files"><a href='../../raise/html/GptThread.html#method-GptThread-upload_files'><code>raise::GptThread$upload_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Quiz" data-id="add_questions"><a href='../../raise/html/Quiz.html#method-Quiz-add_questions'><code>raise::Quiz$add_questions()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Quiz" data-id="add_questions_legacy"><a href='../../raise/html/Quiz.html#method-Quiz-add_questions_legacy'><code>raise::Quiz$add_questions_legacy()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Quiz" data-id="copy_from"><a href='../../raise/html/Quiz.html#method-Quiz-copy_from'><code>raise::Quiz$copy_from()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Quiz" data-id="delete_question"><a href='../../raise/html/Quiz.html#method-Quiz-delete_question'><code>raise::Quiz$delete_question()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Quiz" data-id="export"><a href='../../raise/html/Quiz.html#method-Quiz-export'><code>raise::Quiz$export()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Quiz" data-id="expose"><a href='../../raise/html/Quiz.html#method-Quiz-expose'><code>raise::Quiz$expose()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Quiz" data-id="flush_question"><a href='../../raise/html/Quiz.html#method-Quiz-flush_question'><code>raise::Quiz$flush_question()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Quiz" data-id="get_questions"><a href='../../raise/html/Quiz.html#method-Quiz-get_questions'><code>raise::Quiz$get_questions()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Quiz" data-id="get_questions_md"><a href='../../raise/html/Quiz.html#method-Quiz-get_questions_md'><code>raise::Quiz$get_questions_md()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Quiz" data-id="get_title"><a href='../../raise/html/Quiz.html#method-Quiz-get_title'><code>raise::Quiz$get_title()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Quiz" data-id="power_up"><a href='../../raise/html/Quiz.html#method-Quiz-power_up'><code>raise::Quiz$power_up()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Quiz" data-id="restore"><a href='../../raise/html/Quiz.html#method-Quiz-restore'><code>raise::Quiz$restore()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Quiz" data-id="set_description"><a href='../../raise/html/Quiz.html#method-Quiz-set_description'><code>raise::Quiz$set_description()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Quiz" data-id="set_title"><a href='../../raise/html/Quiz.html#method-Quiz-set_title'><code>raise::Quiz$set_title()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Quiz" data-id="shuffle_choices"><a href='../../raise/html/Quiz.html#method-Quiz-shuffle_choices'><code>raise::Quiz$shuffle_choices()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Quiz" data-id="shuffle_questions"><a href='../../raise/html/Quiz.html#method-Quiz-shuffle_questions'><code>raise::Quiz$shuffle_questions()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuizSlides-new"></a>}}
\if{latex}{\out{\hypertarget{method-QuizSlides-new}{}}}
\subsection{Method \code{new()}}{
QuizSlides class constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuizSlides$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuizSlides-get_description"></a>}}
\if{latex}{\out{\hypertarget{method-QuizSlides-get_description}{}}}
\subsection{Method \code{get_description()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuizSlides$get_description(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuizSlides-clone"></a>}}
\if{latex}{\out{\hypertarget{method-QuizSlides-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuizSlides$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
