% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuizStudy.R
\name{QuizStudy}
\alias{QuizStudy}
\title{QuizStudy class}
\description{
Some description
}
\details{
Some details.
}
\section{Super classes}{
\code{\link[raise:GptThread]{raise::GptThread}} -> \code{\link[raise:Quiz]{raise::Quiz}} -> \code{QuizStudy}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-QuizStudy-new}{\code{QuizStudy$new()}}
\item \href{#method-QuizStudy-get_questions}{\code{QuizStudy$get_questions()}}
\item \href{#method-QuizStudy-get_questions_md}{\code{QuizStudy$get_questions_md()}}
\item \href{#method-QuizStudy-get_description_md}{\code{QuizStudy$get_description_md()}}
\item \href{#method-QuizStudy-get_summary_md}{\code{QuizStudy$get_summary_md()}}
\item \href{#method-QuizStudy-get_description}{\code{QuizStudy$get_description()}}
\item \href{#method-QuizStudy-clone}{\code{QuizStudy$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="chat"><a href='../../raise/html/GptThread.html#method-GptThread-chat'><code>raise::GptThread$chat()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="cover_tracks"><a href='../../raise/html/GptThread.html#method-GptThread-cover_tracks'><code>raise::GptThread$cover_tracks()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="create_vector_store"><a href='../../raise/html/GptThread.html#method-GptThread-create_vector_store'><code>raise::GptThread$create_vector_store()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="get_last_message"><a href='../../raise/html/GptThread.html#method-GptThread-get_last_message'><code>raise::GptThread$get_last_message()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="get_messages"><a href='../../raise/html/GptThread.html#method-GptThread-get_messages'><code>raise::GptThread$get_messages()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="parse_files"><a href='../../raise/html/GptThread.html#method-GptThread-parse_files'><code>raise::GptThread$parse_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="upload_files"><a href='../../raise/html/GptThread.html#method-GptThread-upload_files'><code>raise::GptThread$upload_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Quiz" data-id="add_questions"><a href='../../raise/html/Quiz.html#method-Quiz-add_questions'><code>raise::Quiz$add_questions()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Quiz" data-id="add_questions_legacy"><a href='../../raise/html/Quiz.html#method-Quiz-add_questions_legacy'><code>raise::Quiz$add_questions_legacy()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Quiz" data-id="copy_from"><a href='../../raise/html/Quiz.html#method-Quiz-copy_from'><code>raise::Quiz$copy_from()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Quiz" data-id="delete_question"><a href='../../raise/html/Quiz.html#method-Quiz-delete_question'><code>raise::Quiz$delete_question()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Quiz" data-id="export"><a href='../../raise/html/Quiz.html#method-Quiz-export'><code>raise::Quiz$export()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Quiz" data-id="expose"><a href='../../raise/html/Quiz.html#method-Quiz-expose'><code>raise::Quiz$expose()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Quiz" data-id="flush_question"><a href='../../raise/html/Quiz.html#method-Quiz-flush_question'><code>raise::Quiz$flush_question()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Quiz" data-id="get_title"><a href='../../raise/html/Quiz.html#method-Quiz-get_title'><code>raise::Quiz$get_title()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Quiz" data-id="power_up"><a href='../../raise/html/Quiz.html#method-Quiz-power_up'><code>raise::Quiz$power_up()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Quiz" data-id="restore"><a href='../../raise/html/Quiz.html#method-Quiz-restore'><code>raise::Quiz$restore()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Quiz" data-id="set_description"><a href='../../raise/html/Quiz.html#method-Quiz-set_description'><code>raise::Quiz$set_description()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Quiz" data-id="set_title"><a href='../../raise/html/Quiz.html#method-Quiz-set_title'><code>raise::Quiz$set_title()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Quiz" data-id="shuffle_choices"><a href='../../raise/html/Quiz.html#method-Quiz-shuffle_choices'><code>raise::Quiz$shuffle_choices()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Quiz" data-id="shuffle_questions"><a href='../../raise/html/Quiz.html#method-Quiz-shuffle_questions'><code>raise::Quiz$shuffle_questions()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuizStudy-new"></a>}}
\if{latex}{\out{\hypertarget{method-QuizStudy-new}{}}}
\subsection{Method \code{new()}}{
QuizStudy class constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuizStudy$new(study_design, title, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuizStudy-get_questions"></a>}}
\if{latex}{\out{\hypertarget{method-QuizStudy-get_questions}{}}}
\subsection{Method \code{get_questions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuizStudy$get_questions(include_description = F, heading_level = 1, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuizStudy-get_questions_md"></a>}}
\if{latex}{\out{\hypertarget{method-QuizStudy-get_questions_md}{}}}
\subsection{Method \code{get_questions_md()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuizStudy$get_questions_md(include_description = F, heading_level = 1, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuizStudy-get_description_md"></a>}}
\if{latex}{\out{\hypertarget{method-QuizStudy-get_description_md}{}}}
\subsection{Method \code{get_description_md()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuizStudy$get_description_md(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuizStudy-get_summary_md"></a>}}
\if{latex}{\out{\hypertarget{method-QuizStudy-get_summary_md}{}}}
\subsection{Method \code{get_summary_md()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuizStudy$get_summary_md(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuizStudy-get_description"></a>}}
\if{latex}{\out{\hypertarget{method-QuizStudy-get_description}{}}}
\subsection{Method \code{get_description()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuizStudy$get_description(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuizStudy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-QuizStudy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuizStudy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
