% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study.R, R/StudyBinary.R,
%   R/StudyBinaryPython.R, R/StudyGaussian.R, R/StudyGaussianPython.R,
%   R/helpers.R
\name{Study}
\alias{Study}
\alias{StudyBinary}
\alias{StudyBinaryPython}
\alias{StudyGaussian}
\alias{StudyGaussianPython}
\alias{effect}
\title{Study class}
\usage{
effect(..., effects = c(), custom = NULL)
}
\description{
Some description

Some description

Some description

Some description

Some description
}
\details{
Some details.

Some details.

Some details.

Some details.

Some details.
}
\note{
For debugging only. Don't you dare use this in a package!
}
\section{Super class}{
\code{\link[raise:GptThread]{raise::GptThread}} -> \code{Study}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Study-copy_from}{\code{Study$copy_from()}}
\item \href{#method-Study-get_title}{\code{Study$get_title()}}
\item \href{#method-Study-set_title}{\code{Study$set_title()}}
\item \href{#method-Study-get_link}{\code{Study$get_link()}}
\item \href{#method-Study-get_dist}{\code{Study$get_dist()}}
\item \href{#method-Study-expose}{\code{Study$expose()}}
\item \href{#method-Study-power_up}{\code{Study$power_up()}}
\item \href{#method-Study-new}{\code{Study$new()}}
\item \href{#method-Study-generate_description}{\code{Study$generate_description()}}
\item \href{#method-Study-generate_correlations}{\code{Study$generate_correlations()}}
\item \href{#method-Study-generate_coefficients}{\code{Study$generate_coefficients()}}
\item \href{#method-Study-generate}{\code{Study$generate()}}
\item \href{#method-Study-get_formula}{\code{Study$get_formula()}}
\item \href{#method-Study-get_description}{\code{Study$get_description()}}
\item \href{#method-Study-get_description_json}{\code{Study$get_description_json()}}
\item \href{#method-Study-get_description_md}{\code{Study$get_description_md()}}
\item \href{#method-Study-get_outcome}{\code{Study$get_outcome()}}
\item \href{#method-Study-get_outcome_md}{\code{Study$get_outcome_md()}}
\item \href{#method-Study-get_independent_variables}{\code{Study$get_independent_variables()}}
\item \href{#method-Study-get_independent_variables_md}{\code{Study$get_independent_variables_md()}}
\item \href{#method-Study-get_correlations}{\code{Study$get_correlations()}}
\item \href{#method-Study-get_correlations_json}{\code{Study$get_correlations_json()}}
\item \href{#method-Study-get_coefficients}{\code{Study$get_coefficients()}}
\item \href{#method-Study-get_coefficients_json}{\code{Study$get_coefficients_json()}}
\item \href{#method-Study-get_equation_numbers_md}{\code{Study$get_equation_numbers_md()}}
\item \href{#method-Study-get_equation_labels_md}{\code{Study$get_equation_labels_md()}}
\item \href{#method-Study-sample}{\code{Study$sample()}}
\item \href{#method-Study-get_data}{\code{Study$get_data()}}
\item \href{#method-Study-export}{\code{Study$export()}}
\item \href{#method-Study-get_model}{\code{Study$get_model()}}
\item \href{#method-Study-get_model_summary_md_force}{\code{Study$get_model_summary_md_force()}}
\item \href{#method-Study-get_model_summary_md}{\code{Study$get_model_summary_md()}}
\item \href{#method-Study-get_summary_md}{\code{Study$get_summary_md()}}
\item \href{#method-Study-clone}{\code{Study$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="chat"><a href='../../raise/html/GptThread.html#method-GptThread-chat'><code>raise::GptThread$chat()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="cover_tracks"><a href='../../raise/html/GptThread.html#method-GptThread-cover_tracks'><code>raise::GptThread$cover_tracks()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="create_vector_store"><a href='../../raise/html/GptThread.html#method-GptThread-create_vector_store'><code>raise::GptThread$create_vector_store()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="get_last_message"><a href='../../raise/html/GptThread.html#method-GptThread-get_last_message'><code>raise::GptThread$get_last_message()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="get_messages"><a href='../../raise/html/GptThread.html#method-GptThread-get_messages'><code>raise::GptThread$get_messages()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="parse_files"><a href='../../raise/html/GptThread.html#method-GptThread-parse_files'><code>raise::GptThread$parse_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="restore"><a href='../../raise/html/GptThread.html#method-GptThread-restore'><code>raise::GptThread$restore()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="upload_files"><a href='../../raise/html/GptThread.html#method-GptThread-upload_files'><code>raise::GptThread$upload_files()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-copy_from"></a>}}
\if{latex}{\out{\hypertarget{method-Study-copy_from}{}}}
\subsection{Method \code{copy_from()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$copy_from(from, cache = T)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-get_title"></a>}}
\if{latex}{\out{\hypertarget{method-Study-get_title}{}}}
\subsection{Method \code{get_title()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$get_title()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-set_title"></a>}}
\if{latex}{\out{\hypertarget{method-Study-set_title}{}}}
\subsection{Method \code{set_title()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$set_title(title)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-get_link"></a>}}
\if{latex}{\out{\hypertarget{method-Study-get_link}{}}}
\subsection{Method \code{get_link()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$get_link(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-get_dist"></a>}}
\if{latex}{\out{\hypertarget{method-Study-get_dist}{}}}
\subsection{Method \code{get_dist()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$get_dist(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-expose"></a>}}
\if{latex}{\out{\hypertarget{method-Study-expose}{}}}
\subsection{Method \code{expose()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$expose()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-power_up"></a>}}
\if{latex}{\out{\hypertarget{method-Study-power_up}{}}}
\subsection{Method \code{power_up()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$power_up()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-new"></a>}}
\if{latex}{\out{\hypertarget{method-Study-new}{}}}
\subsection{Method \code{new()}}{
DataGenerator class constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-generate_description"></a>}}
\if{latex}{\out{\hypertarget{method-Study-generate_description}{}}}
\subsection{Method \code{generate_description()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$generate_description(
  title,
  description,
  outcome,
  n_continuous = 0,
  n_categorical = c(),
  n_binary = 0,
  details = "",
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-generate_correlations"></a>}}
\if{latex}{\out{\hypertarget{method-Study-generate_correlations}{}}}
\subsection{Method \code{generate_correlations()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$generate_correlations(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-generate_coefficients"></a>}}
\if{latex}{\out{\hypertarget{method-Study-generate_coefficients}{}}}
\subsection{Method \code{generate_coefficients()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$generate_coefficients(
  effects = list(effect(custom = "all main effects"), effect("categorical", "binary")),
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-generate"></a>}}
\if{latex}{\out{\hypertarget{method-Study-generate}{}}}
\subsection{Method \code{generate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$generate(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-get_formula"></a>}}
\if{latex}{\out{\hypertarget{method-Study-get_formula}{}}}
\subsection{Method \code{get_formula()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$get_formula()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-get_description"></a>}}
\if{latex}{\out{\hypertarget{method-Study-get_description}{}}}
\subsection{Method \code{get_description()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$get_description()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-get_description_json"></a>}}
\if{latex}{\out{\hypertarget{method-Study-get_description_json}{}}}
\subsection{Method \code{get_description_json()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$get_description_json()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-get_description_md"></a>}}
\if{latex}{\out{\hypertarget{method-Study-get_description_md}{}}}
\subsection{Method \code{get_description_md()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$get_description_md(heading_level = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-get_outcome"></a>}}
\if{latex}{\out{\hypertarget{method-Study-get_outcome}{}}}
\subsection{Method \code{get_outcome()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$get_outcome(markdown = F)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-get_outcome_md"></a>}}
\if{latex}{\out{\hypertarget{method-Study-get_outcome_md}{}}}
\subsection{Method \code{get_outcome_md()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$get_outcome_md(markdown = F)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-get_independent_variables"></a>}}
\if{latex}{\out{\hypertarget{method-Study-get_independent_variables}{}}}
\subsection{Method \code{get_independent_variables()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$get_independent_variables(markdown = F)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-get_independent_variables_md"></a>}}
\if{latex}{\out{\hypertarget{method-Study-get_independent_variables_md}{}}}
\subsection{Method \code{get_independent_variables_md()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$get_independent_variables_md()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-get_correlations"></a>}}
\if{latex}{\out{\hypertarget{method-Study-get_correlations}{}}}
\subsection{Method \code{get_correlations()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$get_correlations()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-get_correlations_json"></a>}}
\if{latex}{\out{\hypertarget{method-Study-get_correlations_json}{}}}
\subsection{Method \code{get_correlations_json()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$get_correlations_json()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-get_coefficients"></a>}}
\if{latex}{\out{\hypertarget{method-Study-get_coefficients}{}}}
\subsection{Method \code{get_coefficients()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$get_coefficients(json = F)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-get_coefficients_json"></a>}}
\if{latex}{\out{\hypertarget{method-Study-get_coefficients_json}{}}}
\subsection{Method \code{get_coefficients_json()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$get_coefficients_json()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-get_equation_numbers_md"></a>}}
\if{latex}{\out{\hypertarget{method-Study-get_equation_numbers_md}{}}}
\subsection{Method \code{get_equation_numbers_md()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$get_equation_numbers_md()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-get_equation_labels_md"></a>}}
\if{latex}{\out{\hypertarget{method-Study-get_equation_labels_md}{}}}
\subsection{Method \code{get_equation_labels_md()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$get_equation_labels_md()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-sample"></a>}}
\if{latex}{\out{\hypertarget{method-Study-sample}{}}}
\subsection{Method \code{sample()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$sample(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-Study-get_data}{}}}
\subsection{Method \code{get_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$get_data(centering = F, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-export"></a>}}
\if{latex}{\out{\hypertarget{method-Study-export}{}}}
\subsection{Method \code{export()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$export()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-get_model"></a>}}
\if{latex}{\out{\hypertarget{method-Study-get_model}{}}}
\subsection{Method \code{get_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$get_model(centering = F, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-get_model_summary_md_force"></a>}}
\if{latex}{\out{\hypertarget{method-Study-get_model_summary_md_force}{}}}
\subsection{Method \code{get_model_summary_md_force()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$get_model_summary_md_force()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-get_model_summary_md"></a>}}
\if{latex}{\out{\hypertarget{method-Study-get_model_summary_md}{}}}
\subsection{Method \code{get_model_summary_md()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$get_model_summary_md(force = F, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-get_summary_md"></a>}}
\if{latex}{\out{\hypertarget{method-Study-get_summary_md}{}}}
\subsection{Method \code{get_summary_md()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$get_summary_md(heading_level = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Study-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super classes}{
\code{\link[raise:GptThread]{raise::GptThread}} -> \code{\link[raise:Study]{raise::Study}} -> \code{StudyBinary}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StudyBinary-power_up}{\code{StudyBinary$power_up()}}
\item \href{#method-StudyBinary-new}{\code{StudyBinary$new()}}
\item \href{#method-StudyBinary-generate_description}{\code{StudyBinary$generate_description()}}
\item \href{#method-StudyBinary-export}{\code{StudyBinary$export()}}
\item \href{#method-StudyBinary-get_outcome_md}{\code{StudyBinary$get_outcome_md()}}
\item \href{#method-StudyBinary-get_model}{\code{StudyBinary$get_model()}}
\item \href{#method-StudyBinary-get_model_summary_md_force}{\code{StudyBinary$get_model_summary_md_force()}}
\item \href{#method-StudyBinary-get_equation_numbers_md}{\code{StudyBinary$get_equation_numbers_md()}}
\item \href{#method-StudyBinary-get_equation_labels_md}{\code{StudyBinary$get_equation_labels_md()}}
\item \href{#method-StudyBinary-clone}{\code{StudyBinary$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="chat"><a href='../../raise/html/GptThread.html#method-GptThread-chat'><code>raise::GptThread$chat()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="cover_tracks"><a href='../../raise/html/GptThread.html#method-GptThread-cover_tracks'><code>raise::GptThread$cover_tracks()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="create_vector_store"><a href='../../raise/html/GptThread.html#method-GptThread-create_vector_store'><code>raise::GptThread$create_vector_store()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="get_last_message"><a href='../../raise/html/GptThread.html#method-GptThread-get_last_message'><code>raise::GptThread$get_last_message()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="get_messages"><a href='../../raise/html/GptThread.html#method-GptThread-get_messages'><code>raise::GptThread$get_messages()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="parse_files"><a href='../../raise/html/GptThread.html#method-GptThread-parse_files'><code>raise::GptThread$parse_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="restore"><a href='../../raise/html/GptThread.html#method-GptThread-restore'><code>raise::GptThread$restore()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="upload_files"><a href='../../raise/html/GptThread.html#method-GptThread-upload_files'><code>raise::GptThread$upload_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="copy_from"><a href='../../raise/html/Study.html#method-Study-copy_from'><code>raise::Study$copy_from()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="expose"><a href='../../raise/html/Study.html#method-Study-expose'><code>raise::Study$expose()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="generate"><a href='../../raise/html/Study.html#method-Study-generate'><code>raise::Study$generate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="generate_coefficients"><a href='../../raise/html/Study.html#method-Study-generate_coefficients'><code>raise::Study$generate_coefficients()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="generate_correlations"><a href='../../raise/html/Study.html#method-Study-generate_correlations'><code>raise::Study$generate_correlations()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_coefficients"><a href='../../raise/html/Study.html#method-Study-get_coefficients'><code>raise::Study$get_coefficients()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_coefficients_json"><a href='../../raise/html/Study.html#method-Study-get_coefficients_json'><code>raise::Study$get_coefficients_json()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_correlations"><a href='../../raise/html/Study.html#method-Study-get_correlations'><code>raise::Study$get_correlations()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_correlations_json"><a href='../../raise/html/Study.html#method-Study-get_correlations_json'><code>raise::Study$get_correlations_json()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_data"><a href='../../raise/html/Study.html#method-Study-get_data'><code>raise::Study$get_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_description"><a href='../../raise/html/Study.html#method-Study-get_description'><code>raise::Study$get_description()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_description_json"><a href='../../raise/html/Study.html#method-Study-get_description_json'><code>raise::Study$get_description_json()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_description_md"><a href='../../raise/html/Study.html#method-Study-get_description_md'><code>raise::Study$get_description_md()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_dist"><a href='../../raise/html/Study.html#method-Study-get_dist'><code>raise::Study$get_dist()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_formula"><a href='../../raise/html/Study.html#method-Study-get_formula'><code>raise::Study$get_formula()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_independent_variables"><a href='../../raise/html/Study.html#method-Study-get_independent_variables'><code>raise::Study$get_independent_variables()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_independent_variables_md"><a href='../../raise/html/Study.html#method-Study-get_independent_variables_md'><code>raise::Study$get_independent_variables_md()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_link"><a href='../../raise/html/Study.html#method-Study-get_link'><code>raise::Study$get_link()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_model_summary_md"><a href='../../raise/html/Study.html#method-Study-get_model_summary_md'><code>raise::Study$get_model_summary_md()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_outcome"><a href='../../raise/html/Study.html#method-Study-get_outcome'><code>raise::Study$get_outcome()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_summary_md"><a href='../../raise/html/Study.html#method-Study-get_summary_md'><code>raise::Study$get_summary_md()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_title"><a href='../../raise/html/Study.html#method-Study-get_title'><code>raise::Study$get_title()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="sample"><a href='../../raise/html/Study.html#method-Study-sample'><code>raise::Study$sample()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="set_title"><a href='../../raise/html/Study.html#method-Study-set_title'><code>raise::Study$set_title()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyBinary-power_up"></a>}}
\if{latex}{\out{\hypertarget{method-StudyBinary-power_up}{}}}
\subsection{Method \code{power_up()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyBinary$power_up()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyBinary-new"></a>}}
\if{latex}{\out{\hypertarget{method-StudyBinary-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyBinary$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyBinary-generate_description"></a>}}
\if{latex}{\out{\hypertarget{method-StudyBinary-generate_description}{}}}
\subsection{Method \code{generate_description()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyBinary$generate_description(outcome = "binary", ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyBinary-export"></a>}}
\if{latex}{\out{\hypertarget{method-StudyBinary-export}{}}}
\subsection{Method \code{export()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyBinary$export(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyBinary-get_outcome_md"></a>}}
\if{latex}{\out{\hypertarget{method-StudyBinary-get_outcome_md}{}}}
\subsection{Method \code{get_outcome_md()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyBinary$get_outcome_md()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyBinary-get_model"></a>}}
\if{latex}{\out{\hypertarget{method-StudyBinary-get_model}{}}}
\subsection{Method \code{get_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyBinary$get_model(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyBinary-get_model_summary_md_force"></a>}}
\if{latex}{\out{\hypertarget{method-StudyBinary-get_model_summary_md_force}{}}}
\subsection{Method \code{get_model_summary_md_force()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyBinary$get_model_summary_md_force(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyBinary-get_equation_numbers_md"></a>}}
\if{latex}{\out{\hypertarget{method-StudyBinary-get_equation_numbers_md}{}}}
\subsection{Method \code{get_equation_numbers_md()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyBinary$get_equation_numbers_md(heading_level = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyBinary-get_equation_labels_md"></a>}}
\if{latex}{\out{\hypertarget{method-StudyBinary-get_equation_labels_md}{}}}
\subsection{Method \code{get_equation_labels_md()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyBinary$get_equation_labels_md(heading_level = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyBinary-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StudyBinary-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyBinary$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super classes}{
\code{\link[raise:GptThread]{raise::GptThread}} -> \code{\link[raise:Study]{raise::Study}} -> \code{\link[raise:StudyBinary]{raise::StudyBinary}} -> \code{StudyBinaryPython}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StudyBinaryPython-get_model}{\code{StudyBinaryPython$get_model()}}
\item \href{#method-StudyBinaryPython-get_model_summary_md_force}{\code{StudyBinaryPython$get_model_summary_md_force()}}
\item \href{#method-StudyBinaryPython-get_formula}{\code{StudyBinaryPython$get_formula()}}
\item \href{#method-StudyBinaryPython-get_formula_coding}{\code{StudyBinaryPython$get_formula_coding()}}
\item \href{#method-StudyBinaryPython-clone}{\code{StudyBinaryPython$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="chat"><a href='../../raise/html/GptThread.html#method-GptThread-chat'><code>raise::GptThread$chat()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="cover_tracks"><a href='../../raise/html/GptThread.html#method-GptThread-cover_tracks'><code>raise::GptThread$cover_tracks()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="create_vector_store"><a href='../../raise/html/GptThread.html#method-GptThread-create_vector_store'><code>raise::GptThread$create_vector_store()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="get_last_message"><a href='../../raise/html/GptThread.html#method-GptThread-get_last_message'><code>raise::GptThread$get_last_message()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="get_messages"><a href='../../raise/html/GptThread.html#method-GptThread-get_messages'><code>raise::GptThread$get_messages()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="parse_files"><a href='../../raise/html/GptThread.html#method-GptThread-parse_files'><code>raise::GptThread$parse_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="restore"><a href='../../raise/html/GptThread.html#method-GptThread-restore'><code>raise::GptThread$restore()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="upload_files"><a href='../../raise/html/GptThread.html#method-GptThread-upload_files'><code>raise::GptThread$upload_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="copy_from"><a href='../../raise/html/Study.html#method-Study-copy_from'><code>raise::Study$copy_from()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="expose"><a href='../../raise/html/Study.html#method-Study-expose'><code>raise::Study$expose()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="generate"><a href='../../raise/html/Study.html#method-Study-generate'><code>raise::Study$generate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="generate_coefficients"><a href='../../raise/html/Study.html#method-Study-generate_coefficients'><code>raise::Study$generate_coefficients()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="generate_correlations"><a href='../../raise/html/Study.html#method-Study-generate_correlations'><code>raise::Study$generate_correlations()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_coefficients"><a href='../../raise/html/Study.html#method-Study-get_coefficients'><code>raise::Study$get_coefficients()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_coefficients_json"><a href='../../raise/html/Study.html#method-Study-get_coefficients_json'><code>raise::Study$get_coefficients_json()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_correlations"><a href='../../raise/html/Study.html#method-Study-get_correlations'><code>raise::Study$get_correlations()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_correlations_json"><a href='../../raise/html/Study.html#method-Study-get_correlations_json'><code>raise::Study$get_correlations_json()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_data"><a href='../../raise/html/Study.html#method-Study-get_data'><code>raise::Study$get_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_description"><a href='../../raise/html/Study.html#method-Study-get_description'><code>raise::Study$get_description()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_description_json"><a href='../../raise/html/Study.html#method-Study-get_description_json'><code>raise::Study$get_description_json()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_description_md"><a href='../../raise/html/Study.html#method-Study-get_description_md'><code>raise::Study$get_description_md()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_dist"><a href='../../raise/html/Study.html#method-Study-get_dist'><code>raise::Study$get_dist()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_independent_variables"><a href='../../raise/html/Study.html#method-Study-get_independent_variables'><code>raise::Study$get_independent_variables()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_independent_variables_md"><a href='../../raise/html/Study.html#method-Study-get_independent_variables_md'><code>raise::Study$get_independent_variables_md()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_link"><a href='../../raise/html/Study.html#method-Study-get_link'><code>raise::Study$get_link()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_model_summary_md"><a href='../../raise/html/Study.html#method-Study-get_model_summary_md'><code>raise::Study$get_model_summary_md()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_outcome"><a href='../../raise/html/Study.html#method-Study-get_outcome'><code>raise::Study$get_outcome()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_summary_md"><a href='../../raise/html/Study.html#method-Study-get_summary_md'><code>raise::Study$get_summary_md()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_title"><a href='../../raise/html/Study.html#method-Study-get_title'><code>raise::Study$get_title()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="sample"><a href='../../raise/html/Study.html#method-Study-sample'><code>raise::Study$sample()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="set_title"><a href='../../raise/html/Study.html#method-Study-set_title'><code>raise::Study$set_title()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="StudyBinary" data-id="export"><a href='../../raise/html/StudyBinary.html#method-StudyBinary-export'><code>raise::StudyBinary$export()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="StudyBinary" data-id="generate_description"><a href='../../raise/html/StudyBinary.html#method-StudyBinary-generate_description'><code>raise::StudyBinary$generate_description()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="StudyBinary" data-id="get_equation_labels_md"><a href='../../raise/html/StudyBinary.html#method-StudyBinary-get_equation_labels_md'><code>raise::StudyBinary$get_equation_labels_md()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="StudyBinary" data-id="get_equation_numbers_md"><a href='../../raise/html/StudyBinary.html#method-StudyBinary-get_equation_numbers_md'><code>raise::StudyBinary$get_equation_numbers_md()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="StudyBinary" data-id="get_outcome_md"><a href='../../raise/html/StudyBinary.html#method-StudyBinary-get_outcome_md'><code>raise::StudyBinary$get_outcome_md()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="StudyBinary" data-id="initialize"><a href='../../raise/html/StudyBinary.html#method-StudyBinary-initialize'><code>raise::StudyBinary$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="StudyBinary" data-id="power_up"><a href='../../raise/html/StudyBinary.html#method-StudyBinary-power_up'><code>raise::StudyBinary$power_up()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyBinaryPython-get_model"></a>}}
\if{latex}{\out{\hypertarget{method-StudyBinaryPython-get_model}{}}}
\subsection{Method \code{get_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyBinaryPython$get_model()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyBinaryPython-get_model_summary_md_force"></a>}}
\if{latex}{\out{\hypertarget{method-StudyBinaryPython-get_model_summary_md_force}{}}}
\subsection{Method \code{get_model_summary_md_force()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyBinaryPython$get_model_summary_md_force(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyBinaryPython-get_formula"></a>}}
\if{latex}{\out{\hypertarget{method-StudyBinaryPython-get_formula}{}}}
\subsection{Method \code{get_formula()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyBinaryPython$get_formula()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyBinaryPython-get_formula_coding"></a>}}
\if{latex}{\out{\hypertarget{method-StudyBinaryPython-get_formula_coding}{}}}
\subsection{Method \code{get_formula_coding()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyBinaryPython$get_formula_coding()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyBinaryPython-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StudyBinaryPython-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyBinaryPython$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super classes}{
\code{\link[raise:GptThread]{raise::GptThread}} -> \code{\link[raise:Study]{raise::Study}} -> \code{StudyGaussian}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StudyGaussian-power_up}{\code{StudyGaussian$power_up()}}
\item \href{#method-StudyGaussian-new}{\code{StudyGaussian$new()}}
\item \href{#method-StudyGaussian-get_outcome_md}{\code{StudyGaussian$get_outcome_md()}}
\item \href{#method-StudyGaussian-generate_description}{\code{StudyGaussian$generate_description()}}
\item \href{#method-StudyGaussian-export}{\code{StudyGaussian$export()}}
\item \href{#method-StudyGaussian-get_model}{\code{StudyGaussian$get_model()}}
\item \href{#method-StudyGaussian-get_model_summary_md_force}{\code{StudyGaussian$get_model_summary_md_force()}}
\item \href{#method-StudyGaussian-get_equation_numbers_md}{\code{StudyGaussian$get_equation_numbers_md()}}
\item \href{#method-StudyGaussian-get_equation_labels_md}{\code{StudyGaussian$get_equation_labels_md()}}
\item \href{#method-StudyGaussian-clone}{\code{StudyGaussian$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="chat"><a href='../../raise/html/GptThread.html#method-GptThread-chat'><code>raise::GptThread$chat()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="cover_tracks"><a href='../../raise/html/GptThread.html#method-GptThread-cover_tracks'><code>raise::GptThread$cover_tracks()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="create_vector_store"><a href='../../raise/html/GptThread.html#method-GptThread-create_vector_store'><code>raise::GptThread$create_vector_store()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="get_last_message"><a href='../../raise/html/GptThread.html#method-GptThread-get_last_message'><code>raise::GptThread$get_last_message()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="get_messages"><a href='../../raise/html/GptThread.html#method-GptThread-get_messages'><code>raise::GptThread$get_messages()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="parse_files"><a href='../../raise/html/GptThread.html#method-GptThread-parse_files'><code>raise::GptThread$parse_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="restore"><a href='../../raise/html/GptThread.html#method-GptThread-restore'><code>raise::GptThread$restore()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="upload_files"><a href='../../raise/html/GptThread.html#method-GptThread-upload_files'><code>raise::GptThread$upload_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="copy_from"><a href='../../raise/html/Study.html#method-Study-copy_from'><code>raise::Study$copy_from()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="expose"><a href='../../raise/html/Study.html#method-Study-expose'><code>raise::Study$expose()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="generate"><a href='../../raise/html/Study.html#method-Study-generate'><code>raise::Study$generate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="generate_coefficients"><a href='../../raise/html/Study.html#method-Study-generate_coefficients'><code>raise::Study$generate_coefficients()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="generate_correlations"><a href='../../raise/html/Study.html#method-Study-generate_correlations'><code>raise::Study$generate_correlations()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_coefficients"><a href='../../raise/html/Study.html#method-Study-get_coefficients'><code>raise::Study$get_coefficients()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_coefficients_json"><a href='../../raise/html/Study.html#method-Study-get_coefficients_json'><code>raise::Study$get_coefficients_json()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_correlations"><a href='../../raise/html/Study.html#method-Study-get_correlations'><code>raise::Study$get_correlations()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_correlations_json"><a href='../../raise/html/Study.html#method-Study-get_correlations_json'><code>raise::Study$get_correlations_json()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_data"><a href='../../raise/html/Study.html#method-Study-get_data'><code>raise::Study$get_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_description"><a href='../../raise/html/Study.html#method-Study-get_description'><code>raise::Study$get_description()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_description_json"><a href='../../raise/html/Study.html#method-Study-get_description_json'><code>raise::Study$get_description_json()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_description_md"><a href='../../raise/html/Study.html#method-Study-get_description_md'><code>raise::Study$get_description_md()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_dist"><a href='../../raise/html/Study.html#method-Study-get_dist'><code>raise::Study$get_dist()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_formula"><a href='../../raise/html/Study.html#method-Study-get_formula'><code>raise::Study$get_formula()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_independent_variables"><a href='../../raise/html/Study.html#method-Study-get_independent_variables'><code>raise::Study$get_independent_variables()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_independent_variables_md"><a href='../../raise/html/Study.html#method-Study-get_independent_variables_md'><code>raise::Study$get_independent_variables_md()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_link"><a href='../../raise/html/Study.html#method-Study-get_link'><code>raise::Study$get_link()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_model_summary_md"><a href='../../raise/html/Study.html#method-Study-get_model_summary_md'><code>raise::Study$get_model_summary_md()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_outcome"><a href='../../raise/html/Study.html#method-Study-get_outcome'><code>raise::Study$get_outcome()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_summary_md"><a href='../../raise/html/Study.html#method-Study-get_summary_md'><code>raise::Study$get_summary_md()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_title"><a href='../../raise/html/Study.html#method-Study-get_title'><code>raise::Study$get_title()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="sample"><a href='../../raise/html/Study.html#method-Study-sample'><code>raise::Study$sample()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="set_title"><a href='../../raise/html/Study.html#method-Study-set_title'><code>raise::Study$set_title()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyGaussian-power_up"></a>}}
\if{latex}{\out{\hypertarget{method-StudyGaussian-power_up}{}}}
\subsection{Method \code{power_up()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyGaussian$power_up()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyGaussian-new"></a>}}
\if{latex}{\out{\hypertarget{method-StudyGaussian-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyGaussian$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyGaussian-get_outcome_md"></a>}}
\if{latex}{\out{\hypertarget{method-StudyGaussian-get_outcome_md}{}}}
\subsection{Method \code{get_outcome_md()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyGaussian$get_outcome_md()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyGaussian-generate_description"></a>}}
\if{latex}{\out{\hypertarget{method-StudyGaussian-generate_description}{}}}
\subsection{Method \code{generate_description()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyGaussian$generate_description(outcome = "continuous", ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyGaussian-export"></a>}}
\if{latex}{\out{\hypertarget{method-StudyGaussian-export}{}}}
\subsection{Method \code{export()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyGaussian$export(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyGaussian-get_model"></a>}}
\if{latex}{\out{\hypertarget{method-StudyGaussian-get_model}{}}}
\subsection{Method \code{get_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyGaussian$get_model(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyGaussian-get_model_summary_md_force"></a>}}
\if{latex}{\out{\hypertarget{method-StudyGaussian-get_model_summary_md_force}{}}}
\subsection{Method \code{get_model_summary_md_force()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyGaussian$get_model_summary_md_force(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyGaussian-get_equation_numbers_md"></a>}}
\if{latex}{\out{\hypertarget{method-StudyGaussian-get_equation_numbers_md}{}}}
\subsection{Method \code{get_equation_numbers_md()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyGaussian$get_equation_numbers_md(heading_level = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyGaussian-get_equation_labels_md"></a>}}
\if{latex}{\out{\hypertarget{method-StudyGaussian-get_equation_labels_md}{}}}
\subsection{Method \code{get_equation_labels_md()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyGaussian$get_equation_labels_md(heading_level = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyGaussian-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StudyGaussian-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyGaussian$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super classes}{
\code{\link[raise:GptThread]{raise::GptThread}} -> \code{\link[raise:Study]{raise::Study}} -> \code{\link[raise:StudyGaussian]{raise::StudyGaussian}} -> \code{StudyGaussianPython}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StudyGaussianPython-get_model}{\code{StudyGaussianPython$get_model()}}
\item \href{#method-StudyGaussianPython-get_model_summary_md_force}{\code{StudyGaussianPython$get_model_summary_md_force()}}
\item \href{#method-StudyGaussianPython-get_formula}{\code{StudyGaussianPython$get_formula()}}
\item \href{#method-StudyGaussianPython-get_formula_coding}{\code{StudyGaussianPython$get_formula_coding()}}
\item \href{#method-StudyGaussianPython-clone}{\code{StudyGaussianPython$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="chat"><a href='../../raise/html/GptThread.html#method-GptThread-chat'><code>raise::GptThread$chat()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="cover_tracks"><a href='../../raise/html/GptThread.html#method-GptThread-cover_tracks'><code>raise::GptThread$cover_tracks()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="create_vector_store"><a href='../../raise/html/GptThread.html#method-GptThread-create_vector_store'><code>raise::GptThread$create_vector_store()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="get_last_message"><a href='../../raise/html/GptThread.html#method-GptThread-get_last_message'><code>raise::GptThread$get_last_message()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="get_messages"><a href='../../raise/html/GptThread.html#method-GptThread-get_messages'><code>raise::GptThread$get_messages()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="parse_files"><a href='../../raise/html/GptThread.html#method-GptThread-parse_files'><code>raise::GptThread$parse_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="restore"><a href='../../raise/html/GptThread.html#method-GptThread-restore'><code>raise::GptThread$restore()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="GptThread" data-id="upload_files"><a href='../../raise/html/GptThread.html#method-GptThread-upload_files'><code>raise::GptThread$upload_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="copy_from"><a href='../../raise/html/Study.html#method-Study-copy_from'><code>raise::Study$copy_from()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="expose"><a href='../../raise/html/Study.html#method-Study-expose'><code>raise::Study$expose()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="generate"><a href='../../raise/html/Study.html#method-Study-generate'><code>raise::Study$generate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="generate_coefficients"><a href='../../raise/html/Study.html#method-Study-generate_coefficients'><code>raise::Study$generate_coefficients()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="generate_correlations"><a href='../../raise/html/Study.html#method-Study-generate_correlations'><code>raise::Study$generate_correlations()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_coefficients"><a href='../../raise/html/Study.html#method-Study-get_coefficients'><code>raise::Study$get_coefficients()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_coefficients_json"><a href='../../raise/html/Study.html#method-Study-get_coefficients_json'><code>raise::Study$get_coefficients_json()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_correlations"><a href='../../raise/html/Study.html#method-Study-get_correlations'><code>raise::Study$get_correlations()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_correlations_json"><a href='../../raise/html/Study.html#method-Study-get_correlations_json'><code>raise::Study$get_correlations_json()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_data"><a href='../../raise/html/Study.html#method-Study-get_data'><code>raise::Study$get_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_description"><a href='../../raise/html/Study.html#method-Study-get_description'><code>raise::Study$get_description()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_description_json"><a href='../../raise/html/Study.html#method-Study-get_description_json'><code>raise::Study$get_description_json()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_description_md"><a href='../../raise/html/Study.html#method-Study-get_description_md'><code>raise::Study$get_description_md()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_dist"><a href='../../raise/html/Study.html#method-Study-get_dist'><code>raise::Study$get_dist()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_independent_variables"><a href='../../raise/html/Study.html#method-Study-get_independent_variables'><code>raise::Study$get_independent_variables()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_independent_variables_md"><a href='../../raise/html/Study.html#method-Study-get_independent_variables_md'><code>raise::Study$get_independent_variables_md()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_link"><a href='../../raise/html/Study.html#method-Study-get_link'><code>raise::Study$get_link()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_model_summary_md"><a href='../../raise/html/Study.html#method-Study-get_model_summary_md'><code>raise::Study$get_model_summary_md()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_outcome"><a href='../../raise/html/Study.html#method-Study-get_outcome'><code>raise::Study$get_outcome()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_summary_md"><a href='../../raise/html/Study.html#method-Study-get_summary_md'><code>raise::Study$get_summary_md()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="get_title"><a href='../../raise/html/Study.html#method-Study-get_title'><code>raise::Study$get_title()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="sample"><a href='../../raise/html/Study.html#method-Study-sample'><code>raise::Study$sample()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="Study" data-id="set_title"><a href='../../raise/html/Study.html#method-Study-set_title'><code>raise::Study$set_title()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="StudyGaussian" data-id="export"><a href='../../raise/html/StudyGaussian.html#method-StudyGaussian-export'><code>raise::StudyGaussian$export()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="StudyGaussian" data-id="generate_description"><a href='../../raise/html/StudyGaussian.html#method-StudyGaussian-generate_description'><code>raise::StudyGaussian$generate_description()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="StudyGaussian" data-id="get_equation_labels_md"><a href='../../raise/html/StudyGaussian.html#method-StudyGaussian-get_equation_labels_md'><code>raise::StudyGaussian$get_equation_labels_md()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="StudyGaussian" data-id="get_equation_numbers_md"><a href='../../raise/html/StudyGaussian.html#method-StudyGaussian-get_equation_numbers_md'><code>raise::StudyGaussian$get_equation_numbers_md()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="StudyGaussian" data-id="get_outcome_md"><a href='../../raise/html/StudyGaussian.html#method-StudyGaussian-get_outcome_md'><code>raise::StudyGaussian$get_outcome_md()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="StudyGaussian" data-id="initialize"><a href='../../raise/html/StudyGaussian.html#method-StudyGaussian-initialize'><code>raise::StudyGaussian$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="raise" data-topic="StudyGaussian" data-id="power_up"><a href='../../raise/html/StudyGaussian.html#method-StudyGaussian-power_up'><code>raise::StudyGaussian$power_up()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyGaussianPython-get_model"></a>}}
\if{latex}{\out{\hypertarget{method-StudyGaussianPython-get_model}{}}}
\subsection{Method \code{get_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyGaussianPython$get_model()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyGaussianPython-get_model_summary_md_force"></a>}}
\if{latex}{\out{\hypertarget{method-StudyGaussianPython-get_model_summary_md_force}{}}}
\subsection{Method \code{get_model_summary_md_force()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyGaussianPython$get_model_summary_md_force(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyGaussianPython-get_formula"></a>}}
\if{latex}{\out{\hypertarget{method-StudyGaussianPython-get_formula}{}}}
\subsection{Method \code{get_formula()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyGaussianPython$get_formula()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyGaussianPython-get_formula_coding"></a>}}
\if{latex}{\out{\hypertarget{method-StudyGaussianPython-get_formula_coding}{}}}
\subsection{Method \code{get_formula_coding()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyGaussianPython$get_formula_coding()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyGaussianPython-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StudyGaussianPython-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyGaussianPython$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
