% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ai.R
\name{create_ai}
\alias{create_ai}
\title{Return a tibble based on a prompt}
\usage{
create_ai(prompt, cols, ...)
}
\arguments{
\item{prompt}{The prompt you'd like to run. It could be something like "A spreadsheet of all 50 US states"}

\item{cols}{The columns of the tibble to be returned. Could be something like `c("State", "Area", "Population")`}
}
\description{
You will need an OpenAI API key.
}
\examples{
us_states <- create_ai("A spreadsheet of all 50 US states", c("State", "Area", "Population"))
}
