% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ai.R
\name{extract_ai}
\alias{extract_ai}
\title{Parse a string or vector to a tibble}
\usage{
extract_ai(text, cols, ...)
}
\arguments{
\item{text}{A vector of unstructured text you want to parse to a table.}

\item{cols}{The columns of the tibble to be returned. Could be something like `c("State", "Area", "Population")`}
}
\description{
You will need an OpenAI API key.
}
\examples{
# mps <- read_csv("https://raw.githubusercontent.com/sparkd/mp-financial-interests/master/data/financial-interests-2010-18.csv") |>
  filter(type_code == 1) |>
  head(100) |>
  pluck("description") |>
  extract_ai(cols = c("date", "sum", "donor", "donor_address", "purpose", "hours", "date_registered"))
}
